second commit for 3.6

+++ src/main/java/lang/print/gaps/task6/IntegerTypeCasting.java
@@ -3,7 +3,7 @@ package lang.print.gaps.task6;
 public class IntegerTypeCasting {
     public static void main(String[] args) {
         long first = 1000;
-        int second = (int)first;
+        int second = Math.toIntExact(first);
         System.out.println(first);
         System.out.println(second);
     }
